Add comments to functions
